Add initial Iago registration flow
This has no design consideration, but by going to <url>/initialize, an
iago user can register the app with auth, be brought to <url>finalize,
which tells them what their client ID is, and will save it for them if
the wish that to happen. It is then written to the .iago/config.json
file, though EVERYTHING in the file will be overwritten by the save
(currently).