Add creation of new use case
This adds the necessary pieces of Ember, including controller and
routing, to request a spreadsheet copy and create an ember UseCase
object with the submitted information. It also adds the start of the
express routing to handle saving the new use case.

Still need to implement saving the created use case and spreadsheet id
in a json config file in the .iago folder.