Convert file name spaces to dashes when saving
File name spaces are not converted when a file location is given instead
of a file. Probably need to use some RegEx to get the file name if given
a file location, and convert those spaces to dashes. The problem with
using the entire path is that any path containing spaces would no longer
work.